Handle non-JSON responses from the API
At the moment, the library explodes if it receives a non-JSON
response from the API, with a
`com.google.gson.stream.MalformedJsonException`. This is because
`ResponseParser`'s `parseSingle` method, which all responses are
transformed through (whether successful or an error) assumes
that the body is actually JSON. This may not be the case, for
example when an upstream load balance returns an HTML page in
case of a connectivity problem.

This fixes that by catching any `JsonSyntaxException` and raising
a new `MalformedResponseException`, in line with the other
libraries.

To get there, we also refactor the library read HTTP responses as
strings, not streams (`java.io.Reader`).

This reduces the complexity of code deeper into the library, and
means that we don't have to worry if we ever want to read our body
multiple times, which we want to do in order to be able to include
the errant response body in our `MalformedResponseException`.

There is a trade-off here, in that using a stream enables more
effective memory management. However, this doesn't seem very
important in our library, given that we have control over HTTP
responses and we never sent back a huge amount of content.

However, our string won't have to hang around for long once it is
parsed into JSON (as usually happens), at which point it can be
promptly garbage-collected.